+++ apps/synt-vedtakshistorikk-service/src/main/java/no/nav/testnav/apps/syntvedtakshistorikkservice/consumer/request/search/PersonSearchRequest.java
@@ -1,4 +1,4 @@
-package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.personSearch;
+package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.search;
 
 import lombok.*;
