+++ apps/synt-vedtakshistorikk-service/src/main/java/no/nav/testnav/apps/syntvedtakshistorikkservice/consumer/request/search/PersonstatusSearch.java
@@ -1,4 +1,4 @@
-package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.personSearch;
+package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.search;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,6 @@ import lombok.Value;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(force = true)
-public class AlderSearch {
-    Short fra;
-    Short til;
+public class PersonstatusSearch {
+    String status;
 }
