+++ apps/synt-vedtakshistorikk-service/src/main/java/no/nav/testnav/apps/syntvedtakshistorikkservice/consumer/request/search/RelasjonSearch.java
@@ -1,4 +1,4 @@
-package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.personSearch;
+package no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.search;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,6 +9,6 @@ import lombok.Value;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor(force = true)
-public class PersonstatusSearch {
-    String status;
+public class RelasjonSearch {
+    Boolean barn;
 }
