+++ apps/synt-vedtakshistorikk-service/src/main/java/no/nav/testnav/apps/syntvedtakshistorikkservice/service/IdentService.java
@@ -7,11 +7,11 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.PdlPersonConsumer;
 import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.PersonSearchConsumer;
-import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.personSearch.AlderSearch;
-import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.personSearch.RelasjonSearch;
-import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.personSearch.PersonSearchRequest;
-import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.personSearch.PersonstatusSearch;
-import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.response.PdlPerson;
+import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.search.AlderSearch;
+import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.search.RelasjonSearch;
+import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.search.PersonSearchRequest;
+import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.request.search.PersonstatusSearch;
+import no.nav.testnav.apps.syntvedtakshistorikkservice.consumer.response.pdl.PdlPerson;
 import no.nav.testnav.apps.syntvedtakshistorikkservice.domain.IdentMedKontonr;
 import no.nav.testnav.apps.syntvedtakshistorikkservice.domain.Kontoinfo;
 import no.nav.testnav.libs.dto.personsearchservice.v1.FolkeregisterpersonstatusDTO;
